requirements.txt
opencv-python~=4.6.0.66
numpy~=1.23.1
matplotlib==3.5.2
tqdm~=4.64.0
imgviz~=1.5.0
Pillow~=9.2.0
scipy~=1.8.1
torch~=1.8.2
torchvision~=0.9.2
gpuinfo~=1.0.0a7
pycocotools~=2.0.4